@@ -4,14 +4,14 @@ description: A fun and interactive BMI calculator built with Flutter. Learn the 
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.3 <4.0.0'
+  sdk: '>=3.8.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons:
-  font_awesome_flutter:
+  cupertino_icons: ^1.0.8
+  font_awesome_flutter: ^10.8.0
 
 dev_dependencies:
   flutter_test:
