@@ -4,7 +4,7 @@ import 'package:bmi_calculator/constants.dart';
 class BottomButton extends StatelessWidget {
   BottomButton({required this.onTap, required this.buttonTitle});
 
-  final  onTap;
+  final VoidCallback onTap;
   final String buttonTitle;
 
   @override
