@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
 class ReusableCard extends StatelessWidget {
-  ReusableCard({required this.colour,required this.cardChild,required this.onPress});
+  ReusableCard({required this.colour, required this.cardChild, this.onPress});
 
   final Color colour;
   final Widget cardChild;
-  final  onPress;
+  final VoidCallback? onPress;
 
   @override
   Widget build(BuildContext context) {
