@@ -4,7 +4,7 @@ class RoundIconButton extends StatelessWidget {
   RoundIconButton({required this.icon, required this.onPressed});
 
   final IconData icon;
-  final  onPressed;
+  final VoidCallback onPressed;
 
   @override
   Widget build(BuildContext context) {
